refactor: move color calculation into hook
